[chore] improve folder structure
