--- app/data.py
@@ -1,5 +1,0 @@
-users = [{"id":"3", "name":"Marlone","email":"mar1@gmail.com", "password":"1234","rsvp":True}, {"id":"34", "name":"Pres", "email":"pres911@gmail.com", "password":"3410","rsvp":False}]
-
-events = [{"id":"32", "title":"Rumba", "category":"dance", "location":"nairobi", "description":"A wonderful event extravaganza"}, {"id":"64", "title":"Flamenco", "category":"mariachi", "location":"nakuru", "description":"The best experience so far"}]
-
-logged_users = [{"email":"rian@yahoo.com"}, {"email":"ryu@green.com"}, {"email":"brian@gmail.com"}]
