+++ run.py
@@ -1,0 +1,9 @@
+import os
+
+from app import create_app, db
+
+config_name = os.getenv('FLASK_CONFIG')
+app = create_app(config_name)
+
+if __name__ == '__main__':
+    app.run()
